Tidies up the HMAC-SHA1 signer and removes the base class it extends in preparation to introduce the RSA-SHA1 signer.
